Adding login and signup screens.
