+++ lib/src/features/common/domain/entities/instance.dart
