+++ lib/src/features/common/presentation/widgets/custom_text_field.dart
@@ -1,0 +1,43 @@
+import 'package:flutter/material.dart';
+
+class CustomTextField extends StatelessWidget {
+  final String labelText;
+  final String? errorText;
+  final bool obscureText;
+  final TextInputType textInputType;
+  final Function(String)? onChanged;
+  const CustomTextField({
+    super.key,
+    required this.labelText,
+    this.obscureText = false,
+    this.textInputType = TextInputType.text,
+    this.onChanged,
+    this.errorText,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return TextField(
+      onChanged: onChanged,
+      keyboardType: textInputType,
+      obscureText: obscureText,
+      style:
+          Theme.of(context).textTheme.bodyMedium!.copyWith(color: Colors.black),
+      decoration: InputDecoration(
+        filled: true,
+        fillColor: Colors.white,
+        labelText: labelText,
+        errorText: errorText,
+        floatingLabelBehavior: FloatingLabelBehavior.never,
+        enabledBorder: OutlineInputBorder(
+          borderSide: const BorderSide(color: Colors.black),
+          borderRadius: BorderRadius.circular(10.0),
+        ),
+        focusedBorder: OutlineInputBorder(
+          borderSide: const BorderSide(color: Colors.grey),
+          borderRadius: BorderRadius.circular(10.0),
+        ),
+      ),
+    );
+  }
+}
